Struct google_dfareporting2d2::Account

pub struct Account {
    pub teaser_size_limit: Option<String>,
    pub active_view_opt_out: Option<bool>,
    pub description: Option<String>,
    pub locale: Option<String>,
    pub account_profile: Option<String>,
    pub active: Option<bool>,
    pub comscore_vce_enabled: Option<bool>,
    pub id: Option<String>,
    pub available_permission_ids: Option<Vec<String>>,
    pub kind: Option<String>,
    pub country_id: Option<String>,
    pub name: Option<String>,
    pub maximum_image_size: Option<String>,
    pub nielsen_ocr_enabled: Option<bool>,
    pub default_creative_size_id: Option<String>,
    pub account_permission_ids: Option<Vec<String>>,
    pub currency_id: Option<String>,
    pub reports_configuration: Option<ReportsConfiguration>,
    pub active_ads_limit_tier: Option<String>,
}

Contains properties of a DCM account.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

update accounts (request|response)

list accounts (none)

get accounts (response)

patch accounts (request|response)

Fields
teaser_size_limit: Option<String>

File size limit in kilobytes of Rich Media teaser creatives. Must be between 1 and 10240.

active_view_opt_out: Option<bool>

Whether to serve creatives with Active View tags. If disabled, viewability data will not be available for any impressions.

description: Option<String>

Description of this account.

locale: Option<String>

Locale of this account. Acceptable values are: - "cs" (Czech) - "de" (German) - "en" (English) - "en-GB" (English United Kingdom) - "es" (Spanish) - "fr" (French) - "it" (Italian) - "ja" (Japanese) - "ko" (Korean) - "pl" (Polish) - "pt-BR" (Portuguese Brazil) - "ru" (Russian) - "sv" (Swedish) - "tr" (Turkish) - "zh-CN" (Chinese Simplified) - "zh-TW" (Chinese Traditional)

account_profile: Option<String>

Profile for this account. This is a read-only field that can be left blank.

active: Option<bool>

Whether this account is active.

comscore_vce_enabled: Option<bool>

Whether campaigns created in this account will be enabled for comScore vCE by default.

id: Option<String>

ID of this account. This is a read-only, auto-generated field.

available_permission_ids: Option<Vec<String>>

User role permissions available to the user roles of this account.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#account".

country_id: Option<String>

ID of the country associated with this account.

name: Option<String>

Name of this account. This is a required field, and must be less than 128 characters long and be globally unique.

maximum_image_size: Option<String>

Maximum image size allowed for this account.

nielsen_ocr_enabled: Option<bool>

Whether campaigns created in this account will be enabled for Nielsen OCR reach ratings by default.

default_creative_size_id: Option<String>

Default placement dimensions for this account.

account_permission_ids: Option<Vec<String>>

Account permissions assigned to this account.

currency_id: Option<String>

ID of currency associated with this account. This is a required field. Acceptable values are: - "1" for USD - "2" for GBP - "3" for ESP - "4" for SEK - "5" for CAD - "6" for JPY - "7" for DEM - "8" for AUD - "9" for FRF - "10" for ITL - "11" for DKK - "12" for NOK - "13" for FIM - "14" for ZAR - "15" for IEP - "16" for NLG - "17" for EUR - "18" for KRW - "19" for TWD - "20" for SGD - "21" for CNY - "22" for HKD - "23" for NZD - "24" for MYR - "25" for BRL - "26" for PTE - "27" for MXP - "28" for CLP - "29" for TRY - "30" for ARS - "31" for PEN - "32" for ILS - "33" for CHF - "34" for VEF - "35" for COP - "36" for GTQ - "37" for PLN - "39" for INR - "40" for THB - "41" for IDR - "42" for CZK - "43" for RON - "44" for HUF - "45" for RUB - "46" for AED - "47" for BGN - "48" for HRK

reports_configuration: Option<ReportsConfiguration>

Reporting configuration of this account.

active_ads_limit_tier: Option<String>

Maximum number of active ads allowed for this account.