Struct google_dfareporting2d2::Activities

pub struct Activities {
    pub kind: Option<String>,
    pub filters: Option<Vec<DimensionValue>>,
    pub metric_names: Option<Vec<String>>,
}

Represents an activity group.

This type is not used in any activity, and only used as part of another schema.

Fields

kind

The kind of resource this is, in this case dfareporting#activities.

filters

List of activity filters. The dimension values need to be all either of type "dfa:activity" or "dfa:activityGroup".

metric_names

List of names of floodlight activity metrics.