Struct google_dfareporting2d2::Advertiser

pub struct Advertiser {
    pub status: Option<String>,
    pub kind: Option<String>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub default_click_through_event_tag_id: Option<String>,
    pub original_floodlight_configuration_id: Option<String>,
    pub click_through_url_suffix: Option<String>,
    pub floodlight_configuration_id: Option<String>,
    pub advertiser_group_id: Option<String>,
    pub floodlight_configuration_id_dimension_value: Option<DimensionValue>,
    pub suspended: Option<bool>,
    pub id_dimension_value: Option<DimensionValue>,
    pub default_email: Option<String>,
    pub id: Option<String>,
    pub account_id: Option<String>,
}

Contains properties of a DCM advertiser.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get advertisers (response)

update advertisers (request|response)

patch advertisers (request|response)

list advertisers (none)

insert advertisers (request|response)

Fields

status

Status of this advertiser.

kind

Identifies what kind of resource this is. Value: the fixed string "dfareporting#advertiser".

subaccount_id

Subaccount ID of this advertiser.This is a read-only field that can be left blank.

name

Name of this advertiser. This is a required field and must be less than 256 characters long and unique among advertisers of the same account.

default_click_through_event_tag_id

ID of the click-through event tag to apply by default to the landing pages of this advertiser's campaigns.

original_floodlight_configuration_id

Original floodlight configuration before any sharing occurred. Set the floodlightConfigurationId of this advertiser to originalFloodlightConfigurationId to unshare the advertiser's current floodlight configuration. You cannot unshare an advertiser's floodlight configuration if the shared configuration has activities associated with any campaign or placement.

click_through_url_suffix

Suffix added to click-through URL of ad creative associations under this advertiser. Must be less than 129 characters long.

floodlight_configuration_id

Floodlight configuration ID of this advertiser. The floodlight configuration ID will be created automatically, so on insert this field should be left blank. This field can be set to another advertiser's floodlight configuration ID in order to share that advertiser's floodlight configuration with this advertiser, so long as: - This advertiser's original floodlight configuration is not already associated with floodlight activities or floodlight activity groups. - This advertiser's original floodlight configuration is not already shared with another advertiser.

advertiser_group_id

ID of the advertiser group this advertiser belongs to. You can group advertisers for reporting purposes, allowing you to see aggregated information for all advertisers in each group.

floodlight_configuration_id_dimension_value

Dimension value for the ID of the floodlight configuration. This is a read-only, auto-generated field.

suspended

Suspension status of this advertiser.

id_dimension_value

Dimension value for the ID of this advertiser. This is a read-only, auto-generated field.

default_email

Default email address used in sender field for tag emails.

id

ID of this advertiser. This is a read-only, auto-generated field.

account_id

Account ID of this advertiser.This is a read-only field that can be left blank.