Struct google_dfareporting2d2::Campaign

pub struct Campaign {
    pub comment: Option<String>,
    pub start_date: Option<String>,
    pub end_date: Option<String>,
    pub click_through_url_suffix_properties: Option<ClickThroughUrlSuffixProperties>,
    pub archived: Option<bool>,
    pub additional_creative_optimization_configurations: Option<Vec<CreativeOptimizationConfiguration>>,
    pub advertiser_id: Option<String>,
    pub audience_segment_groups: Option<Vec<AudienceSegmentGroup>>,
    pub create_info: Option<LastModifiedInfo>,
    pub default_click_through_event_tag_properties: Option<DefaultClickThroughEventTagProperties>,
    pub advertiser_id_dimension_value: Option<DimensionValue>,
    pub event_tag_overrides: Option<Vec<EventTagOverride>>,
    pub creative_optimization_configuration: Option<CreativeOptimizationConfiguration>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub trafficker_emails: Option<Vec<String>>,
    pub creative_group_ids: Option<Vec<String>>,
    pub subaccount_id: Option<String>,
    pub name: Option<String>,
    pub lookback_configuration: Option<LookbackConfiguration>,
    pub nielsen_ocr_enabled: Option<bool>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub comscore_vce_enabled: Option<bool>,
    pub kind: Option<String>,
    pub external_id: Option<String>,
    pub advertiser_group_id: Option<String>,
    pub id_dimension_value: Option<DimensionValue>,
    pub billing_invoice_code: Option<String>,
}

Contains properties of a DCM campaign.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

insert campaigns (request|response)

update campaigns (request|response)

patch campaigns (request|response)

get campaigns (response)

list campaigns (none)

Fields
comment: Option<String>

Arbitrary comments about this campaign. Must be less than 256 characters long.

start_date: Option<String>

Date on which the campaign starts running. The start date can be any date. The hours, minutes, and seconds of the start date should not be set, as doing so will result in an error. This is a required field.

end_date: Option<String>

Date on which the campaign will stop running. On insert, the end date must be today or a future date. The end date must be later than or be the same as the start date. If, for example, you set 6/25/2015 as both the start and end dates, the effective campaign run date is just that day only, 6/25/2015. The hours, minutes, and seconds of the end date should not be set, as doing so will result in an error. This is a required field.

click_through_url_suffix_properties: Option<ClickThroughUrlSuffixProperties>

Click-through URL suffix override properties for this campaign.

archived: Option<bool>

Whether this campaign has been archived.

additional_creative_optimization_configurations: Option<Vec<CreativeOptimizationConfiguration>>

Additional creative optimization configurations for the campaign.

advertiser_id: Option<String>

Advertiser ID of this campaign. This is a required field.

audience_segment_groups: Option<Vec<AudienceSegmentGroup>>

Audience segment groups assigned to this campaign. Cannot have more than 300 segment groups.

create_info: Option<LastModifiedInfo>

Information about the creation of this campaign. This is a read-only field.

default_click_through_event_tag_properties: Option<DefaultClickThroughEventTagProperties>

Click-through event tag ID override properties for this campaign.

advertiser_id_dimension_value: Option<DimensionValue>

Dimension value for the advertiser ID of this campaign. This is a read-only, auto-generated field.

event_tag_overrides: Option<Vec<EventTagOverride>>

Overrides that can be used to activate or deactivate advertiser event tags.

creative_optimization_configuration: Option<CreativeOptimizationConfiguration>

Creative optimization configuration for the campaign.

id: Option<String>

ID of this campaign. This is a read-only auto-generated field.

account_id: Option<String>

Account ID of this campaign. This is a read-only field that can be left blank.

trafficker_emails: Option<Vec<String>>

Campaign trafficker contact emails.

creative_group_ids: Option<Vec<String>>

List of creative group IDs that are assigned to the campaign.

subaccount_id: Option<String>

Subaccount ID of this campaign. This is a read-only field that can be left blank.

name: Option<String>

Name of this campaign. This is a required field and must be less than 256 characters long and unique among campaigns of the same advertiser.

lookback_configuration: Option<LookbackConfiguration>

Lookback window settings for the campaign.

nielsen_ocr_enabled: Option<bool>

Whether Nielsen reports are enabled for this campaign.

last_modified_info: Option<LastModifiedInfo>

Information about the most recent modification of this campaign. This is a read-only field.

comscore_vce_enabled: Option<bool>

Whether comScore vCE reports are enabled for this campaign.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#campaign".

external_id: Option<String>

External ID for this campaign.

advertiser_group_id: Option<String>

Advertiser group ID of the associated advertiser.

id_dimension_value: Option<DimensionValue>

Dimension value for the ID of this campaign. This is a read-only, auto-generated field.

billing_invoice_code: Option<String>

Billing invoice code included in the DCM client billing invoices associated with the campaign.