Struct google_dfareporting2d2::DfpSettings

pub struct DfpSettings {
    pub dfp_network_name: Option<String>,
    pub dfp_network_code: Option<String>,
    pub publisher_portal_only: Option<bool>,
    pub programmatic_placement_accepted: Option<bool>,
    pub pub_paid_placement_accepted: Option<bool>,
}

DFP Settings

This type is not used in any activity, and only used as part of another schema.

Fields
dfp_network_name: Option<String>

DFP network name for this directory site.

dfp_network_code: Option<String>

DFP network code for this directory site.

publisher_portal_only: Option<bool>

Whether this directory site is available only via DoubleClick Publisher Portal.

programmatic_placement_accepted: Option<bool>

Whether this directory site accepts programmatic placements.

pub_paid_placement_accepted: Option<bool>

Whether this directory site accepts publisher-paid tags.