Struct google_dfareporting2d2::FloodlightActivityListCall

pub struct FloodlightActivityListCall<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

Retrieves a list of floodlight activities, possibly filtered.

A builder for the list method supported by a floodlightActivity resource. It is not used directly, but through a FloodlightActivityMethods instance.

Example

Instantiate a resource method builder

 
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.floodlight_activities().list("profileId")
             .tag_string("eirmod")
             .sort_order("sadipscing")
             .sort_field("Lorem")
             .search_string("sed")
             .page_token("sit")
             .max_results(-33)
             .add_ids("et")
             .floodlight_configuration_id("At")
             .floodlight_activity_group_type("sit")
             .floodlight_activity_group_tag_string("ut")
             .floodlight_activity_group_name("diam")
             .add_floodlight_activity_group_ids("tempor")
             .advertiser_id("et")
             .doit();

Methods

impl<'a, C, A> FloodlightActivityListCall<'a, C, A> where C: BorrowMut<Client>, A: GetToken

fn doit(self) -> Result<(Response, FloodlightActivitiesListResponse)>

Perform the operation you have build so far.

fn profile_id(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

User profile ID associated with this request.

Sets the profile id path property to the given value.

Even though the property as already been set when instantiating this call, we provide this method for API completeness.

fn tag_string(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Select only floodlight activities with the specified tag string.

Sets the tag string query property to the given value.

fn sort_order(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Order of sorted results, default is ASCENDING.

Sets the sort order query property to the given value.

fn sort_field(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Field by which to sort the list.

Sets the sort field query property to the given value.

fn search_string(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Allows searching for objects by name or ID. Wildcards (*) are allowed. For example, "floodlightactivity*2015" will return objects with names like "floodlightactivity June 2015", "floodlightactivity April 2015", or simply "floodlightactivity 2015". Most of the searches also add wildcards implicitly at the start and the end of the search string. For example, a search string of "floodlightactivity" will match objects with name "my floodlightactivity activity", "floodlightactivity 2015", or simply "floodlightactivity".

Sets the search string query property to the given value.

fn page_token(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Value of the nextPageToken from the previous result page.

Sets the page token query property to the given value.

fn max_results(self, new_value: i32) -> FloodlightActivityListCall<'a, C, A>

Maximum number of results to return.

Sets the max results query property to the given value.

fn add_ids(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Select only floodlight activities with the specified IDs. Must specify either ids, advertiserId, or floodlightConfigurationId for a non-empty result.

Append the given value to the ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn floodlight_configuration_id(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Select only floodlight activities for the specified floodlight configuration ID. Must specify either ids, advertiserId, or floodlightConfigurationId for a non-empty result.

Sets the floodlight configuration id query property to the given value.

fn floodlight_activity_group_type(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Select only floodlight activities with the specified floodlight activity group type.

Sets the floodlight activity group type query property to the given value.

fn floodlight_activity_group_tag_string(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Select only floodlight activities with the specified floodlight activity group tag string.

Sets the floodlight activity group tag string query property to the given value.

fn floodlight_activity_group_name(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Select only floodlight activities with the specified floodlight activity group name.

Sets the floodlight activity group name query property to the given value.

fn add_floodlight_activity_group_ids(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Select only floodlight activities with the specified floodlight activity group IDs.

Append the given value to the floodlight activity group ids query property. Each appended value will retain its original ordering and be '/'-separated in the URL's parameters.

fn advertiser_id(self, new_value: &str) -> FloodlightActivityListCall<'a, C, A>

Select only floodlight activities for the specified advertiser ID. Must specify either ids, advertiserId, or floodlightConfigurationId for a non-empty result.

Sets the advertiser id query property to the given value.

fn delegate(self, new_value: &'a mut Delegate) -> FloodlightActivityListCall<'a, C, A>

The delegate implementation is consulted whenever there is an intermediate result, or if something goes wrong while executing the actual API request.

It should be used to handle progress information, and to implement a certain level of resilience.

Sets the delegate property to the given value.

fn param<T>(self, name: T, value: T) -> FloodlightActivityListCall<'a, C, A> where T: AsRef<str>

Set any additional parameter of the query string used in the request. It should be used to set parameters which are not yet available through their own setters.

Please note that this method must not be used to set any of the known paramters which have their own setter method. If done anyway, the request will fail.

Additional Parameters

quotaUser (query-string) - Available to use for quota purposes for server-side applications. Can be any arbitrary string assigned to a user, but should not exceed 40 characters. Overrides userIp if both are provided.

oauth_token (query-string) - OAuth 2.0 token for the current user.

key (query-string) - API key. Your API key identifies your project and provides you with API access, quota, and reports. Required unless you provide an OAuth 2.0 token.

prettyPrint (query-boolean) - Returns response with indentations and line breaks.

userIp (query-string) - IP address of the site where the request originates. Use this if you want to enforce per-user limits.

fields (query-string) - Selector specifying which fields to include in a partial response.

alt (query-string) - Data format for the response.

fn add_scope<T>(self, scope: T) -> FloodlightActivityListCall<'a, C, A> where T: AsRef<str>

Identifies the authorization scope for the method you are building.

Use this method to actively specify which scope should be used, instead the default Scope variant Scope::Dfatrafficking.

The scope will be added to a set of scopes. This is important as one can maintain access tokens for more than one scope.

Usually there is more than one suitable scope to authorize an operation, some of which may encompass more rights than others. For example, for listing resources, a read-only scope will be sufficient, a read-write scope will do as well.