Struct google_dfareporting2d2::OrderMethods

pub struct OrderMethods<'a, C, A> where C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on order resources. It is not used directly, but through the Dfareporting hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate yup_oauth2 as oauth2;
extern crate google_dfareporting2d2 as dfareporting2d2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use dfareporting2d2::Dfareporting;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Dfareporting::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *CallBuilders*
// like `get(...)` and `list(...)`
// to build up your call.
let rb = hub.orders();

Methods

impl<'a, C, A> OrderMethods<'a, C, A>

fn list(&self, profile_id: &str, project_id: &str) -> OrderListCall<'a, C, A>

Create a builder to help you perform the following task:

Retrieves a list of orders, possibly filtered.

Arguments

profileId - User profile ID associated with this request.

projectId - Project ID for orders.

fn get(&self, profile_id: &str, project_id: &str, id: &str) -> OrderGetCall<'a, C, A>

Create a builder to help you perform the following task:

Gets one order by ID.

Arguments

profileId - User profile ID associated with this request.

projectId - Project ID for orders.

id - Order ID.