Struct google_dfareporting2d2::Project

pub struct Project {
    pub start_date: Option<String>,
    pub target_cpm_nanos: Option<String>,
    pub end_date: Option<String>,
    pub client_billing_code: Option<String>,
    pub overview: Option<String>,
    pub target_clicks: Option<String>,
    pub target_cpa_nanos: Option<String>,
    pub advertiser_id: Option<String>,
    pub target_impressions: Option<String>,
    pub subaccount_id: Option<String>,
    pub audience_gender: Option<String>,
    pub id: Option<String>,
    pub account_id: Option<String>,
    pub kind: Option<String>,
    pub target_cpc_nanos: Option<String>,
    pub name: Option<String>,
    pub last_modified_info: Option<LastModifiedInfo>,
    pub budget: Option<String>,
    pub audience_age_group: Option<String>,
    pub target_conversions: Option<String>,
    pub client_name: Option<String>,
}

Contains properties of a DoubleClick Planning project.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

get projects (response)

list projects (none)

Fields
start_date: Option<String>

Start date of the project.

target_cpm_nanos: Option<String>

CPM that the advertiser is targeting.

end_date: Option<String>

End date of the project.

client_billing_code: Option<String>

Client billing code of this project.

overview: Option<String>

Overview of this project.

target_clicks: Option<String>

Number of clicks that the advertiser is targeting.

target_cpa_nanos: Option<String>

CPA that the advertiser is targeting.

advertiser_id: Option<String>

Advertiser ID of this project.

target_impressions: Option<String>

Number of impressions that the advertiser is targeting.

subaccount_id: Option<String>

Subaccount ID of this project.

audience_gender: Option<String>

Audience gender of this project.

id: Option<String>

ID of this project. This is a read-only, auto-generated field.

account_id: Option<String>

Account ID of this project.

kind: Option<String>

Identifies what kind of resource this is. Value: the fixed string "dfareporting#project".

target_cpc_nanos: Option<String>

CPC that the advertiser is targeting.

name: Option<String>

Name of this project.

last_modified_info: Option<LastModifiedInfo>

Information about the most recent modification of this project.

budget: Option<String>

Budget of this project in the currency specified by the current account. The value stored in this field represents only the non-fractional amount. For example, for USD, the smallest value that can be represented by this field is 1 US dollar.

audience_age_group: Option<String>

Audience age group of this project.

target_conversions: Option<String>

Number of conversions that the advertiser is targeting.

client_name: Option<String>

Name of the project client.