Struct google_dfareporting2d2::ReportPathToConversionCriteriaReportProperties

pub struct ReportPathToConversionCriteriaReportProperties {
    pub clicks_lookback_window: Option<i32>,
    pub pivot_on_interaction_path: Option<bool>,
    pub impressions_lookback_window: Option<i32>,
    pub include_unattributed_ip_conversions: Option<bool>,
    pub include_unattributed_cookie_conversions: Option<bool>,
    pub maximum_interaction_gap: Option<i32>,
    pub include_attributed_ip_conversions: Option<bool>,
    pub maximum_click_interactions: Option<i32>,
    pub maximum_impression_interactions: Option<i32>,
}

The properties of the report.

This type is not used in any activity, and only used as part of another schema.

Fields

clicks_lookback_window

DFA checks to see if a click interaction occurred within the specified period of time before a conversion. By default the value is pulled from Floodlight or you can manually enter a custom value. Valid values: 1-90.

pivot_on_interaction_path

Enable pivoting on interaction path.

impressions_lookback_window

DFA checks to see if an impression interaction occurred within the specified period of time before a conversion. By default the value is pulled from Floodlight or you can manually enter a custom value. Valid values: 1-90.

include_unattributed_ip_conversions

Include conversions that have no associated cookies and no exposures. It’s therefore impossible to know how the user was exposed to your ads during the lookback window prior to a conversion.

include_unattributed_cookie_conversions

Include conversions of users with a DoubleClick cookie but without an exposure. That means the user did not click or see an ad from the advertiser within the Floodlight group, or that the interaction happened outside the lookback window.

maximum_interaction_gap

The maximum amount of time that can take place between interactions (clicks or impressions) by the same user. Valid values: 1-90.

include_attributed_ip_conversions

Deprecated: has no effect.

maximum_click_interactions

The maximum number of click interactions to include in the report. Advertisers currently paying for E2C reports get up to 200 (100 clicks, 100 impressions). If another advertiser in your network is paying for E2C, you can have up to 5 total exposures per report.

maximum_impression_interactions

The maximum number of click interactions to include in the report. Advertisers currently paying for E2C reports get up to 200 (100 clicks, 100 impressions). If another advertiser in your network is paying for E2C, you can have up to 5 total exposures per report.