Struct google_dfareporting2d2::UserDefinedVariableConfiguration

pub struct UserDefinedVariableConfiguration {
    pub data_type: Option<String>,
    pub report_name: Option<String>,
    pub variable_type: Option<String>,
}

User Defined Variable configuration.

This type is not used in any activity, and only used as part of another schema.

Fields
data_type: Option<String>

Data type for the variable. This is a required field.

report_name: Option<String>

User-friendly name for the variable which will appear in reports. This is a required field, must be less than 64 characters long, and cannot contain the following characters: ""<>".

variable_type: Option<String>

Variable name in the tag. This is a required field.